+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 - Change `--json` global arg to now be `--log-file FILE`
 - Add `--mode lines` and `--mode auto` to `ct pick` command
 - Change `PickerTui` to no longer be generic over `<T>`, instead the `pick` and `pick_many` fns own the generic
+- Fix `ct pick` using stdout instead of stderr when restoring from Ratatui
+- Fix `ct pick` default query not being valid after changing default engine to `jmes-path`
+- Add `ct pick --engine liquid` for using [liquid](https://github.com/cobalt-org/liquid-rust)
 
 # v0.30.0